+++ src/EmailException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Wavey\LaravelGravatar;
+
+class EmailException extends \Exception
+{
+    // TODO Add some stuffs... maybe... in here.
+}